requirements.txt
# ── Web scraping ───────────────────────────────────────────────────────────────
# app-store-scraper removed: hard-pins requests==2.23.0 (conflicts with modern stack)
# We call Apple's public iTunes Search + RSS APIs directly via httpx instead.
httpx>=0.27.0
google-play-scraper>=1.2.4

# ── Embeddings (ONNX Runtime — prebuilt cp314 wheels for Windows) ──────────────
# Runs BAAI/bge-small-en-v1.5 locally via ONNX, no PyTorch required.
onnxruntime>=1.18.0
transformers>=4.40.0
tokenizers>=0.19.0
huggingface-hub>=0.23.0

# ── Numeric ────────────────────────────────────────────────────────────────────
numpy>=2.0.0

# ── Vector store ──────────────────────────────────────────────────────────────
chromadb>=0.5.0

# ── LLM client (OpenAI-compatible, used for OpenRouter) ───────────────────────
openai>=1.30.0

# ── Sentiment analysis (pure-Python, no native extensions) ────────────────────
vaderSentiment>=3.3.2

# ── UI ─────────────────────────────────────────────────────────────────────────
streamlit>=1.35.0
plotly>=5.20.0

# ── Config ─────────────────────────────────────────────────────────────────────
python-dotenv>=1.0.0
